+++ CHANGELOG.md
@@ -8,6 +8,10 @@ CHANGELOG
 
 ---
 
+* 0.2.6 (2013-06-01)
+
+ * Guzzle 3.6 support
+
 * 0.2.5 (2013-04-01)
 
  * Fixed Hixie-76 handshake bug